+++ tests/api/DXFeedTest.cpp
@@ -5,7 +5,6 @@
 
 #include <string>
 #include <thread>
-#include <unordered_map>
 #include <vector>
 
 #include <dxfeed_graal_c_api/api.h>
@@ -18,7 +17,7 @@ using namespace dxfcpp::literals;
 using namespace std::literals;
 
 TEST_CASE("DXFeed::getIndexedEventsIfSubscribed should return an empty vector when not subscribed to any events") {
-    auto events = DXFeed::getInstance()->getIndexedEventsIfSubscribed<Order>("AAPL", OrderSource::ntv);
+    const auto events = DXFeed::getInstance()->getIndexedEventsIfSubscribed<Order>("AAPL", OrderSource::ntv);
 
     REQUIRE(events.empty());
 }
@@ -28,14 +27,18 @@ struct DXFeedTestFixture {
     DXEndpoint::Ptr endpoint{};
     DXFeed::Ptr feed{};
     DXPublisher::Ptr pub{};
+    std::shared_ptr<InPlaceExecutor> executor{};
+    std::vector<std::shared_ptr<EventType>> publishedEvents{};
 
     DXFeedTestFixture() {
+        executor = InPlaceExecutor::create();
         endpoint = DXEndpoint::newBuilder()
                        ->withRole(DXEndpoint::Role::LOCAL_HUB)
                        ->withProperty(DXEndpoint::DXFEED_WILDCARD_ENABLE_PROPERTY, "true")
                        ->withProperty(DXEndpoint::DXENDPOINT_EVENT_TIME_PROPERTY, "true")
                        ->withProperty(DXEndpoint::DXSCHEME_NANO_TIME_PROPERTY, "true")
                        ->build();
+        endpoint->executor(executor);
         feed = endpoint->getFeed();
         pub = endpoint->getPublisher();
     }
@@ -51,6 +54,36 @@ struct DXFeedTestFixture {
             .sharedAs<Order>();
     }
 
+    std::shared_ptr<Candle> createCandle(std::int64_t time) const {
+        return std::make_shared<Candle>(CandleSymbol::valueOf(symbol))
+            ->withTime(time).withClose(42.0).withVolume(1000)
+            .sharedAs<Candle>();
+    }
+
+    void publish(const auto &event) {
+        publishedEvents.push_back(event);
+    }
+
+    void publishMany(const auto &events) {
+        publishedEvents.insert(publishedEvents.end(), std::begin(events), std::end(events));
+    }
+
+    void process() {
+        pub->publishEvents(publishedEvents);
+        executor->processAllPendingTasks();
+        publishedEvents.clear();
+    }
+
+    void publishAndProcess(const auto &event) {
+        publish(event);
+        process();
+    }
+
+    void publishAndProcessMany(const auto &events) {
+        publishMany(events);
+        process();
+    }
+
     ~DXFeedTestFixture() {
         endpoint->awaitProcessed();
         endpoint->close();
@@ -58,14 +91,40 @@ struct DXFeedTestFixture {
 };
 
 TEST_CASE_FIXTURE(DXFeedTestFixture, "DXFeed::getIndexedEventsIfSubscribed should return events") {
-    auto o = createOrder(0, Side::BUY, 1, 1,
-                         static_cast<std::int32_t>((EventFlag::SNAPSHOT_BEGIN | EventFlag::SNAPSHOT_END).getMask()));
+    const auto o = createOrder(
+        0, Side::BUY, 1, 1, static_cast<std::int32_t>((EventFlag::SNAPSHOT_BEGIN | EventFlag::SNAPSHOT_END).getMask()));
+    const auto sub = feed->createSubscription(Order::TYPE);
 
-    auto sub = feed->createSubscription(Order::TYPE);
     sub->addSymbols(symbol);
+    publishAndProcess(o);
 
-    pub->publishEvents(o);
-    auto events = feed->getIndexedEventsIfSubscribed<Order>(symbol, OrderSource::DEFAULT);
+    const auto events = feed->getIndexedEventsIfSubscribed<Order>(symbol, OrderSource::DEFAULT);
 
     REQUIRE(events.size() == 1);
+}
+
+TEST_CASE("DXFeed::getTimeSeriesIfSubscribed should return an empty vector when not subscribed to any events") {
+    const auto fromTime = std::chrono::milliseconds(now()) - std::chrono::days(1);
+    constexpr auto toTime = std::chrono::milliseconds(std::numeric_limits<std::int64_t>::max());
+    const auto events =
+        DXFeed::getInstance()->getTimeSeriesIfSubscribed<Candle>("AAPL", fromTime.count(), toTime.count());
+
+    REQUIRE(events.empty());
+}
+
+TEST_CASE_FIXTURE(DXFeedTestFixture, "DXFeed::getTimeSeriesIfSubscribed should return events") {
+    const auto fromTime = std::chrono::milliseconds(now()) - std::chrono::days(1);
+    const auto c = createCandle(fromTime.count());
+    const auto sub = feed->createSubscription(Candle::TYPE);
+
+    sub->addSymbols(TimeSeriesSubscriptionSymbol(symbol, fromTime.count()));
+
+    publishAndProcess(c);
+
+    const auto events = feed->getTimeSeriesIfSubscribed<Candle>(symbol, fromTime.count());
+
+    REQUIRE(events.size() == 1);
+    REQUIRE(events[0]->getTime() == fromTime.count());
+    REQUIRE(events[0]->getClose() == doctest::Approx(c->getClose()));
+    REQUIRE(events[0]->getVolume() == doctest::Approx(c->getVolume()));
 }